// app/types.ts
export enum TipoDocumento {
  CPF = 'CPF',
  RG = 'RG',
  CNH = 'CNH',
  PASSAPORTE = 'PASSAPORTE',
  RNE = 'RNE',
  CNPJ = 'CNPJ',
  OUTRO = 'OUTRO'
}